% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/regrets.R
\name{regrets_photo}
\alias{regrets_photo}
\title{Save a random egret photo to file}
\usage{
regrets_photo(file = "test.jpg")
}
\arguments{
\item{file}{path where you want to save your egret .jpg picture}
}
\description{
Save a random egret photo to file
}
